### 2020-09-17 #2682 @whedon
[PRE REVIEW]: Bempp-cl: A fast Python based just-in-time compiling boundary element library.
**Submitting author:** @mscroggs (<a href="http://orcid.org/0000-0002-4658-2443">Matthew Scroggs</a>)
**Repository:** <a href="https://github.com/bempp/bempp-cl" target ="_blank">https://github.com/bempp/bempp-cl</a>
**Version:** v0.2.1
