### 2020-09-17 #2682 @kyleniemeyer
Hi @meg-simula, can you edit this submission?

@mscroggs we'll be using this issue to get things started, including finding reviewers—any suggestions you have are welcome.
