### 2020-09-19 9de6d94
Add workflow to automatically check that instances of version numbers agree
