+++ .github/workflows/on-push.yml
@@ -70,6 +70,7 @@ jobs:
       - run: | 
             python3 -m black --line-length 88 --check test
         name: Run black checks on tests
+
   build-and-test-with-exafmm:
     name: Build and run tests with ExaFMM
     needs: build-and-publish-docker
@@ -82,3 +83,13 @@ jobs:
         name: Install Bempp
       - run: python3 -m pytest --durations=50 test/unit
         name: Run Bempp unit tests
+
+  check-version-number:
+    name: Check that all version numbers agree
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v2
+      - name: Install PyGitHub
+        run: pip3 install PyGitHub
+      - name: Update website
+        run: python3 .github/scripts/update_version_number.py ${{ secrets.bemppbot_github_token }} ${GITHUB_REF##*/}
