+++ bempp/version.py
@@ -1,1 +1,1 @@
-__version__ = "0.2.2"
+__version__ = "0.2.0"
