### 2020-09-19 a46978b
Build docker image with version tag when version is created
