+++ .github/workflows/on-release.yml
@@ -48,6 +48,6 @@ jobs:
         run: |
           echo ::set-output name=SOURCE_TAG::${GITHUB_REF#refs/tags/}
       - name: Install twitter api
-        steps: pip3 install twitter
+        run: pip3 install twitter
       - name: Tweet
         run: python3 .github/scripts/tweet.py ${{ steps.tag_name.outputs.SOURCE_TAG }} ${{ secrets.twitter1 }} ${{ secrets.twitter2 }} ${{ secrets.twitter3 }} ${{ secrets.twitter4 }}
