### 2020-09-21 #2682 @meg-simula
@mscroggs Thanks for the submission, nice work! Before proceeding, could I ask you to take a look at the JOSS submission guidelines regarding paper length, and recommended structure:

https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain
