+++ joss/paper.bib
@@ -43,19 +43,6 @@ AUTHOR = {S. M. Rao and D. R. Wilton and A. W. Glisson},
     doi = {10.1109/TAP.1982.1142818}
 }
 
-@article{rt,
-    AUTHOR = {Raviart, Pierre-Arnaud and Thomas, Jean-Marie},
-    EDITOR = {Galligani, Ilio and Magenes, Enrico},
-     TITLE = {A mixed finite element method for 2nd order elliptic problems},
- BOOKTITLE = {Mathematical aspects of finite element methods},
-     PAGES = {292--315},
-    SERIES = {Lecture Notes in Mathematics},
-    VOLUME = {606},
-      YEAR = {1977},
- PUBLISHER = {Springer},
-    DOI = {10.1007/BFb0064470}
-}
-
 @article{nedelec,
     AUTHOR = {N\'ed\'elec, Jean-Claude},
      TITLE = {Mixed finite elements in {$\mathbb{R}^3$}},
@@ -172,13 +159,4 @@ AUTHOR = {S. M. Rao and D. R. Wilton and A. W. Glisson},
     NUMBER = {2},
      PAGES = {6:1--6:40},
     DOI = {10.1145/2590830}
-}
-
-@book{ColtonKress,
-     TITLE = {Integral Equation Methods in Scattering Theory},
-    AUTHOR = {Colton, David and Kress, Rainer},
-      YEAR = {1983},
- PUBLISHER = {Wiley},
-     PAGES = {xv+270},
-    DOI = {10.1002/zamm.19850650104}
 }
