### 2020-10-05 f978527
Shorten JOSS paper
