+++ joss/paper.md
@@ -24,20 +24,26 @@ bibliography: paper.bib
 
 # Summary
 
-The boundary element method (BEM) is a numerical method for approximating the solution of certain types of partial differential equations (PDEs) in homogeneous bounded or unbounded domains.
-The method finds the approximation by discretising a boundary integral equation that can be derived from the PDE. The mathematical
-background of BEM is covered in, for example, @Stein07 or @McLean. Typical applications of BEM include electrostatic problems, and acoustic and electromagnetic scattering.
+The boundary element method (BEM) is a numerical method for approximating the solution of certain types of partial 
+differential equations (PDEs) in homogeneous bounded or unbounded domains. The method finds the approximation by discretising 
+a boundary integral equation that can be derived from the PDE. The mathematical background of BEM is covered in, for example, 
+@Stein07 or @McLean. Typical applications of BEM include electrostatic problems, and acoustic and electromagnetic scattering.
 
 Bempp-cl is an open-source boundary element method library that can be used to assemble all the standard integral kernels for
 Laplace, Helmholtz, modified Helmholtz, and Maxwell problems. The library has a user-friendly Python interface that allows the
 user to use BEM to solve a variety of problems, including problems in electrostatics, acoustics and electromagnetics.
 
-Bempp-cl began life as BEM++, and was a Python library with a C++ computational core. The ++ slowly changed into pp as
-functionality gradually moved from C++ to Python with only a few core routines remaining in C++. Bempp-cl is the culmination of efforts to fully move to Python. It is an almost complete rewrite of Bempp: the C++ core has been replaced by highly SIMD optimised just-in-time compiled OpenCL kernels, or alternatively, by just-in-time compiled Numba routines, which are automatically used on systems that do not provide OpenCL drivers. User visible functionality is strictly separated from the implementation of computational routines, making it easy to add other discretisation technologies in the future (e.g. future support for SYCL-based heterogeneous compute devices).
+Bempp-cl began life as BEM++, and was a Python library with a C++ computational core. The ++ slowly changed into pp as 
+functionality gradually moved from C++ to Python with only a few core routines remaining in C++. Bempp-cl is the culmination 
+of efforts to fully move to Python, and is an almost complete rewrite of Bempp/
 
 In this paper, we give an overview of the functionality of Bempp-cl and present highlights of the library's recent developments.
 An overview of the original version of the library is presented in @bemppold. Full documentation of the library can be found
 online at ``bempp.com`` and in @bempphandbook.
+
+# Statement of need
+
+
 
 # An overview of Bempp features
 
@@ -57,59 +63,52 @@ There are five main steps that are commonly taken when solving a problem with BE
 Bempp-cl provides routines that implement each of these steps.
 
 ## Grid Interface
-The submodule `bempp.api.shapes` contains the definitions of a number of shapes. From these, grids with various element
-sizes can be created internally using Gmsh [@gmsh]. Alternatively, meshes can be imported from many formats using the
-meshio library [@meshio]. Bempp-cl currently only supports flat triangle based surface meshes. Higher-order triangular meshes may be supported
-in the future.
+The submodule `bempp.api.shapes` contains the definitions of a number of shapes. From these, grids with various element sizes 
+can be created internally using Gmsh [@gmsh]. Alternatively, meshes can be imported from many formats using the meshio library 
+[@meshio]. Bempp-cl currently only supports flat triangle based surface meshes. Higher-order triangular meshes may be 
+supported in the future.
 
 ## Function Spaces
[…]
