### 2020-10-05 14b0fb1
Update paper.md
Small updates
