### 2020-10-12 029b056
Add test that local2global and global2local agree
