### 2020-10-12 138f3b2
run tests on screen and cubes
