### 2020-10-12 0d20675
raise valuerror is BC space construction attempted on screen
