+++ bempp/api/space/space.py
@@ -431,6 +431,10 @@ class FunctionSpace(object):
         """Return order of the space."""
         return self._order
 
+    def cell_dofs(self, cell_index):
+        """Return the DOF numbers associated with the cell."""
+        dofs = [-1 if j == 0 else i for i, j in zip(self.local2global[cell_index], self.local_multipliers)]
+
     @property
     def local2global(self):
         """Return local to global map."""
