+++ bempp/api/space/space.py
@@ -433,7 +433,12 @@ class FunctionSpace(object):
 
     def cell_dofs(self, cell_index):
         """Return the DOF numbers associated with the cell."""
-        dofs = [-1 if j == 0 else i for i, j in zip(self.local2global[cell_index], self.local_multipliers)]
+        return [
+            -1 if j == 0 else i
+            for i, j in zip(
+                self.local2global[cell_index], self.local_multipliers[cell_index]
+            )
+        ]
 
     @property
     def local2global(self):
