+++ bempp/api/space/maxwell_spaces.py
@@ -307,7 +307,7 @@ def bc_function_space(
     """Define a space of BC functions."""
     from .space import SpaceBuilder
 
-    if len(grid.vertices[0]) - len(grid.edges[0]) + len(grid.elements[0]):
+    if len(grid.vertices[0]) - len(grid.edges[0]) + len(grid.elements[0]) != 2:
         # Grid is a screen, not a polyhedron
         raise ValueError("BC spaces not yet supported on screens")
 
@@ -359,7 +359,7 @@ def rbc_function_space(
     """Define a space of RBC functions."""
     from .space import SpaceBuilder
 
-    if len(grid.vertices[0]) - len(grid.edges[0]) + len(grid.elements[0]):
+    if len(grid.vertices[0]) - len(grid.edges[0]) + len(grid.elements[0]) != 2:
         # Grid is a screen, not a polyhedron
         raise ValueError("BC spaces not yet supported on screens")
 
