### 2020-10-12 e4e9552
use np.close, not == 0
