### 2020-10-12 66d94c7
use None in test, not -1
