+++ test/unit/space/test_local2global.py
@@ -59,4 +59,4 @@ def test_local2global(
         for j, dof in enumerate(globals):
             if dof == -1:
                 assert _np.isclose(space.local_multipliers[i][j], 0)
-        assert (globals == space.cell_dofs(i)).all()
+                assert space.cell_dofs(i)[j] is None
