### 2020-10-12 58798f8
Add test of function io
