### 2020-10-13 afd480d
Refactored mass matrix generation.
