### 2020-10-13 60cec23
Wrapped deletion of FMM temp files in try/except block.
