+++ bempp/api/assembly/discrete_boundary_operator.py
@@ -436,6 +436,10 @@ class ZeroDiscreteBoundaryOperator(_DiscreteOperatorBase):
 
     def _matmat(self, x):
         return _np.zeros((self.shape[0], x.shape[1]), dtype="float64")
+    @property
+    def A(self):
+        """Return as dense."""
+        return _np.zeros((self.shape[0], self.shape[1]), dtype="float64")
 
 
 class DiscreteRankOneOperator(_DiscreteOperatorBase):
