+++ .github/workflows/on-push.yml
@@ -1,5 +1,6 @@
 on:
   push
+  pull_request
 
 jobs:
   build-and-publish-docker:
