### 2020-10-14 d7298e7
currect syntax for running on both
