+++ .github/workflows/on-push.yml
@@ -1,6 +1,4 @@
-on:
-  push
-  pull_request
+on: [push, pull_request]
 
 jobs:
   build-and-publish-docker:
