### 2020-10-15 9f51d66
sparse ZeroDiscreteBoundaryOperator
