### 2020-10-26 554b9c7
A property for SparseDiscreteBoundaryOperator and ZeroDiscreteBoundaryOperator
