+++ bempp/api/assembly/blocked_operator.py
@@ -688,11 +688,17 @@ class BlockedDiscreteOperator(_DiscreteOperatorBase):
     @property
     def A(self):
         """TODO: add docstring."""
+        from bempp.api.assembly.discrete_boundary_operator import SparseDiscreteBoundaryOperator, ZeroDiscreteBoundaryOperator
+
         rows = []
         for i in range(self._ndims[0]):
             row = []
             for j in range(self._ndims[1]):
-                row.append(self[i, j].A)
+                op = self[i,j]
+                temp = op.A
+                if isinstance(op,SparseDiscreteBoundaryOperator) or isinstance(op,ZeroDiscreteBoundaryOperator):
+                    temp = temp.A
+                row.append(temp)
             rows.append(_np.hstack(row))
         return _np.vstack(rows)
 
