### 2020-10-27 a4a1a57
update CI config
