+++ .github/workflows/on-pr.yml
@@ -1,0 +1,50 @@
+on: pull-request
+
+jobs:
+  style-checks:
+    name: Run style checks
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v2
+      - run: sudo apt-get install -y python3-setuptools
+      - run: pip3 install black>=20.8
+      - run: pip3 install pydocstyle flake8
+      - run: |
+            python3 -m flake8 bempp
+            python3 -m flake8 test
+        name: Run flake8 checks
+      - run: python3 -m pydocstyle bempp/api
+        name: Run pydocstyle checks
+      - run: | 
+            python3 -m black --line-length 88 --check bempp
+        name: Run black checks in bempp
+      - run: | 
+            python3 -m black --line-length 88 --check test
+        name: Run black checks on tests
+      - run: | 
+            python3 -m black --line-length 88 --check .github/scripts
+        name: Run black checks on CI scripts
+
+  build-and-test-with-exafmm:
+    name: Build and run tests with ExaFMM
+    runs-on: ubuntu-latest
+    container:
+      image: bempp/cl-dev-env:with-exafmm
+    steps:
+      - uses: actions/checkout@v2
+      - run: python3 setup.py install
+        name: Install Bempp
+      - run: python3 -m pytest --durations=50 test/unit
+        name: Run Bempp unit tests
+
+  check-version-number:
+    name: Check version numbers
+    runs-on: ubuntu-20.04
+    steps:
+      - uses: actions/checkout@v2
+      - name: Install PyGitHub
+        run: pip3 install PyGitHub
+      - name: Debug branch name
+        run: echo ${GITHUB_REF#refs/heads/}
+      - name: Check and update version numbers
+        run: python3 .github/scripts/update_version_number.py ${{ secrets.bemppbot_github_token }} ${GITHUB_REF#refs/heads/}
