+++ .github/workflows/on-push.yml
@@ -46,7 +46,7 @@ jobs:
           username: ${{ secrets.DOCKER_USERNAME }}
           password: ${{ secrets.DOCKER_PASSWORD }}
           repository: bempp/cl-dev-env
-          tags: with-exafmm
+          tags: latest
           target: bempp-dev-env-with-exafmm
           cache_froms: bempp/cl-dev-env
         if: steps.diff.outputs.changed == 'yes'
