+++ .github/workflows/on-push.yml
@@ -1,6 +1,24 @@
 on: push
 
 jobs:
+  get-branch:
+    name: Get branch name
+    runs-on: ubuntu-latest
+    outputs:
+      tag: ${{ steps.vars.outputs.tag }}
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v2
+      - name: Get tag name
+        id: vars
+        run: |
+            if [ ${GITHUB_REF#refs/*/} == "master" ]
+            then
+              echo ::set-output name=tag::latest
+            else
+              echo ::set-output name=tag::${GITHUB_REF#refs/*/}
+            fi
+
   check-version-number:
     name: Check version numbers
     runs-on: ubuntu-20.04
@@ -16,6 +34,7 @@ jobs:
   build-and-publish-docker:
     name: Build and publish testing environment docker image
     runs-on: ubuntu-latest
+    needs: get-branch
     steps:
       - name: Checkout code
         uses: actions/checkout@v2
@@ -46,7 +65,7 @@ jobs:
           username: ${{ secrets.DOCKER_USERNAME }}
           password: ${{ secrets.DOCKER_PASSWORD }}
           repository: bempp/cl-dev-env
-          tags: latest
+          tags: ${{ needs.get-branch.outputs.tag }}
           target: bempp-dev-env-with-exafmm
           cache_froms: bempp/cl-dev-env
         if: steps.diff.outputs.changed == 'yes'
@@ -77,10 +96,11 @@ jobs:
 
   build-and-test-with-exafmm:
     name: Build and run tests with ExaFMM
-    needs: build-and-publish-docker
+    needs:
+      - build-and-publish-docker
+      - get-branch
     runs-on: ubuntu-latest
-    container:
-      image: bempp/cl-dev-env:with-exafmm
+    container: bempp/cl-dev-env:${{ needs.get-branch.outputs.tag }}
     steps:
       - uses: actions/checkout@v2
       - run: python3 setup.py install
