+++ .github/workflows/on-master-push.yml
@@ -5,7 +5,7 @@ on:
 
 jobs:
   build-and-publish-docker:
-    name: Build and publish testing environment docker image
+    name: Build and publish notebook docker image
     runs-on: ubuntu-latest
     steps:
       - name: Checkout code
@@ -19,42 +19,3 @@ jobs:
           tags: latest
           target: lab
           cache_froms: bempp/cl-notebook:latest
-
-  build-and-publish-docker:
-    name: Build and publish testing environment docker image
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout code
-        uses: actions/checkout@v2
-      - name: Make cache folder
-        run: mkdir ~/cache
-      - name: Cache Dockerfile
-        id: cache-docker
-        uses: actions/cache@v2
-        with:
-          path: ~/cache/Dockerfile
-          key: ${{ runner.os }}-Dockerfile-${{ hashFiles('Dockerfile') }}
-      - name: touch ~/cache/Dockerfile
-        run: touch ~/cache/Dockerfile
-      - name: compare Dockerfile with cache
-        id: diff
-        run: |
-            if diff Dockerfile ~/cache/Dockerfile > /dev/null; then
-                echo ::set-output name=changed::no
-            else
-                echo ::set-output name=changed::yes
-            fi
-      - name: Copy Dockerfile to ~/cache/Dockerfile
-        run: cp Dockerfile ~/cache/Dockerfile
-
-      - name: Build and push testing Docker image with exafmm
-        uses: docker/build-push-action@v1
-        with:
-          username: ${{ secrets.DOCKER_USERNAME }}
-          password: ${{ secrets.DOCKER_PASSWORD }}
-          repository: bempp/cl-dev-env
-          tags: latest
-          target: bempp-dev-env-with-exafmm
-          cache_froms: bempp/cl-dev-env
-        if: steps.diff.outputs.changed == 'yes'
-
