+++ .github/workflows/on-push.yml
@@ -65,7 +65,7 @@ jobs:
           username: ${{ secrets.DOCKER_USERNAME }}
           password: ${{ secrets.DOCKER_PASSWORD }}
           repository: bempp/cl-dev-env
-          tags: ${{ needs.get-branch.outputs.tag }}
+          tags: ${{ needs.get-branch.outputs.tag//\/_ }}
           target: bempp-dev-env-with-exafmm
           cache_froms: bempp/cl-dev-env
         if: steps.diff.outputs.changed == 'yes'
@@ -100,7 +100,7 @@ jobs:
       - build-and-publish-docker
       - get-branch
     runs-on: ubuntu-latest
-    container: bempp/cl-dev-env:${{ needs.get-branch.outputs.tag }}
+    container: bempp/cl-dev-env:${{ needs.get-branch.outputs.tag//\/_ }}
     steps:
       - uses: actions/checkout@v2
       - run: python3 setup.py install
