### 2020-12-11 #2879 @meg-simula
@ramisetti: I see. Hm. Did you try these points?

>
> If you cannot edit the checklist please:
>
> Make sure you're logged in to your GitHub account
> Be sure to accept the invite at this URL: https://github.com/openjournals/joss-reviews/invitations
