### 2020-12-11 fd47882
added multitrace_sphere
