+++ bempp/api/shapes/__init__.py
@@ -11,3 +11,7 @@ from .shapes import almond
 from .shapes import reentrant_cube
 from .shapes import screen
 from .shapes import cylinders
+from .shapes import multitrace_ellipsoid
+from .shapes import multitrace_sphere
+
+
