+++ bempp/api/shapes/__init__.py
@@ -13,5 +13,3 @@ from .shapes import screen
 from .shapes import cylinders
 from .shapes import multitrace_ellipsoid
 from .shapes import multitrace_sphere
-
-
