### 2020-12-11 bempp/bempp-cl#89 @jamtrott
Review of Bempp-cl for JOSS
Dear Matthew Scroggs and Timo Betcke,

Here you will find my review of the Bempp-cl package for the Journal
of Open Source Software. The review is also tracked at
https://github.com/openjournals/joss-reviews/issues/2879.

Overview
---------

Bempp-cl is a Python package for solving boundary integral equations
on unstructured meshes for a selection of boundary integral operators
and function spaces. The assembly of linear systems from these
operators can employ either hand-written OpenCL kernels or
just-in-time compilation of Python and numpy code using numba.

The package also provides an interface to linear solvers from the
scipy package, including a direct solver based on LU factorisation and
iterative solvers using the conjugate gradient (CG) method or GMRES.

The Bempp-cl package may optionally interface with FEniCS for finite
element discretisations. It can also use exafmm-t to employ fast
multipole methods (FMMs).

Comments
--------

I have organised my comments according to the review checklist
provided by JOSS.

Conflict of interest

1. I confirm that I have read the JOSS conflict of interest (COI)
   policy and that: I have no COIs with reviewing this work or that
   any perceived COIs have been waived by JOSS for the purpose of this
   review.

Code of Conduct

1. I confirm that I read and will adhere to the JOSS code of conduct.

General checks

1. The repository is available at https://github.com/bempp/bempp-cl/.

2. The code is licensed according to an MIT license, described in the
   file LICENSE.md. Note that JOSS requires the license file to be
   plain text, and it must be named either LICENSE or COPYING. Please
   consider renaming the file accordingly.

3. The authors, Matthew Scroggs and Timo Betcke, are the main authors
   of the code. They have both contributed significantly.

4. The authors' contribution represents a substantial effort in
   providing a high-level Python library for solving boundary integral
   equations.

Functionality

1. I was able to install the package from source, as described in the
   package documentation. Dependencies are clearly described. However,
   due to not having an OpenCL runtime available on the machine I am
   using, I intalled without OpenCL support. The package falls back to
   using numba in this case, as described in the documentation.

2. I confirm that the software can indeed be used to assemble linear
   systems of equations based on standard operators from boundary
   integral equations, including Laplace, Helmholtz, modified
   Helmholtz and Maxwell problems. Both OpenCL and numba kernels are
   implemented, and I have verified that numba may be used for
   just-in-time compilation of the operator assembly kernels.

   Following the tutorial problems in the documentation, I can verify
   that a Laplace problem with Dirichlet boundary conditions is solved
   quite easily through the high-level Python API. Similarly, I can
   verify that more advanced problems may be solved by following
   further tutorials given in the documentation. This includes, for
   example, problems based on Maxwell's equations concerning the
   scattering of an electromagnetic wave from a collection of
   perfectly conducting screens.

3. The software makes no particular claims about performance.

Documentation

1. The package's README and other documentation clearly states the
   need for the software and its scope. There is no explicit mention
   of a target audience, but presumably this would be researchers and
   others who are interested in solving boundary integral equations.

2. The installation instructions are sufficient for installing from
   source, and the dependencies are made clear.

3. The documentation includes several tutorials, ranging from basic
   to more advanced problems, showing how to use the package.

4. A high-level overview of the API is given, and each component
   (e.g., grids, function spaces, operators, and so on) is briefly
   described, sometimes together with code examples.

   In addition, the documentation links to an API reference, which can
   be found at https://bempp-cl.readthedocs.io/en/latest/. In the code
   itself and in the API reference, most functions have been given at
   least basic documentation, describing their purpose. A few
   functions are thoroughly documented with a description of their
   parameters and return value.

5. There appears to be a fairly comprehensive suite of unit
   tests. However, there are no instructions for how to run them. I
   was able to do so with the command

   `$ pytest test/'

   Note that some FEniCS- and FMM-related tests failed, since those
   (optional) dependencies were not installed. This is perhaps
   expected, but it might be better to skip these tests instead of
   failing.

   The tutorials can be considered another form of manual verification
   through which the software can be tested.

6. There are no guidelines for contributing to the software. Users are
   encouraged to submit issues to the GitHub issue tracker to report
   problems, or to contact the developers on Discourse to get support.

Software paper

1. A high-level description of the purpose of the software is given,
   though a background in boundary element methods is assumed. Readers
   are pointed to relevant literature for mathematical background. But
   I think the paper could benefit from a little more background, such
   as an example of the type of equations that are relevant.

   That being said, the overview of Bempp features is clear and
   well-written.

2. The authors state that the software is intended for electrostatic,
   acoustic and electromagnetic scattering problems. Again, there is
   no explicit mention of a target audience, but it is reasonable to
   assume that the audience is researchers working on such problems.

   The statement of need could benefit from a short discussion or
   mention of why they choose to use OpenCL and numba for just-in-time
   compilation. Presumably, this is a feature that sets the software
   apart from other packages, and it is good to inform the reader why
   such a choice was made.

3. There is mention of an earlier version of the same package, but
   there is lack of comparison to other software for solving boundary
   integral equations.

4. The paper is well written and appropriately structured. There are a few minor typos.

   In the Section "Summary": "The method finds the approximation"
      should perhaps be "The method finds an approximation".
   In the Section "Statement of need": Calder'on -> Calderón
   In the Section "Discretisation and solvers": writting -> written
   In the Section "Potential and far field operators": fiels -> field
   In the Section "Further information": in in -> in

5. The DOI for Steinbach (2008) is incorrect, or, at least, following the link
   does not work. Otherwise, the citations appear fine.
