### 2020-12-12 #2879 @ramisetti
@meg-simula: I tried as you suggested but I still cannot edit the checklist page.
