### 2020-12-12 #2879 @ramisetti
Dear @meg-simula,

As I am unable to edit the review checklist, I am providing my review here. Please read below for my review of the _Bempp-cl: A fast Python based just-in-time compiling boundary element library_ package for the **Journal of Open Source Software**.
Conflict of interest
•	I confirm that I have read the JOSS conflict of interest (COI) policy and that: I have no COIs with reviewing this work or that any perceived COIs have been waived by JOSS for the purpose of this review.
Code of Conduct
•	I confirm that I read and will adhere to the JOSS code of conduct.
General checks
•	Repository: Is the source code for this software available at the repository url? Yes
•	License: Does the repository contain a plain-text LICENSE file with the contents of an OSI approved software license? The repository contains LICENSE.md file with the contents of MIT license. The authors should consider renaming it to either LICENSE or COPYING as required for JOSS.
•	Contribution and authorship: Has the submitting author (@mscroggs) made major contributions to the software? Does the full list of paper authors seem appropriate and complete? The submitting authors have made major contributions to the software and the list of the authors seem appropriate and complete as well.
•	Substantial scholarly effort: Does this submission meet the scope eligibility described in the JOSS guidelines. This submission meets the scope eligibility as per the JOSS guidelines and it is clearly visible in terms of the length of commit history, number of commits, LOC, and its potential to attract researchers in this domain.
Functionality
•	 Installation: Does installation proceed as outlined in the documentation? I tried installing the software following instructions for the Quick Install and Installing from source. Both the approaches works smoothly without any issues with Python-3.8 on Mac OS. I have neither tried running it using Docker nor with OpenCL as I have don’t have them.
•	Functionality: Have the functional claims of the software been confirmed? The documentation claims the software can import and export in a number of formats, including Gmsh and VTK but it is not clearly evident from the examples provided in the applications section, except the exporting of gmsh file in the second example. The authors should provide at least one example for each case. Similarly, the authors should provide examples on the CPU and GPU parallelisation. The documentation for the examples should place a phrase related to the functionality to help the users understand the purpose of the example. For example, the documentation for each example can mention a phrase about the functionality in brackets beside the link.
•	Performance: If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.) The authors make no performance claims.
Documentation
•	 A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is? The authors clearly state the type of problems that can be solved using the software. The authors may consider mentioning about the target audience of this software to make it more clear to the potential users
•	 Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution. The documentation mentions the list of required software to install the software. The dependencies are required to be explicitly installed by the user prior to installing the software. The software lacks the automated package management solution. The authors may consider mentioning the versions of the dependencies particularly for gmsh as this could save time for users in installing the dependencies.
•	Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems). The documentation includes good number of examples on how to use the software. The authors should fix the broken link for “Weakly imposing a Dirichlet boundary condition” example. I would recommend to separate the tutorials from documentation as it is more intuitive to find examples in tutorials section rather in documentation
•	Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)? The python core functionality is available, but I believe it is not up to a satisfactory level to be useful for the prospective software developers interested in the software
•	 Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified? Although the repository contains a list of unit test but the documentation does not mention anything about any automated tests. The authors should consider updating the documentation with information on unit tests.
•	 Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support: Information on third party contribution to the software is not provided. The documentation provides information on how to reporting issues or problems through GitHub repository. It also provides an email to seek any support related to the software.
Software paper
•	 Summary: Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided? The paper tries to give a high-level functionality description through good references to BEM. I believe this will help non-specialist readers as well.
•	 A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is? The authors clearly state the type of problems that can be solved using the software. The authors may consider mentioning about the target audience of this software to make it more clear to readers
•	 State of the field: Do the authors describe how this software compares to other commonly-used packages? This submission lacks information on why this software is better and how this compares to any of the existing open source software. The authors should consider including information in the paper if there are any existing other opensource software and how this software compares with them.
•	Quality of writing: Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)? The paper is well written and structured as per the JOSS guidelines. However, there are few minor mistakes that should be corrected to make it better. i) “Calder’on” should be changed to “Calderón” in Section: Statement of need ii) “writing” should be changed to “written” in Section: Discretisation and solvers, iii) “fiels” should be changed to “field” in Section: Potential and far field operators, and iv) “in in” should be changed to “in” in Section: Further information
•	References: Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper citation syntax? I think the list of references is complete and everything is cited appropriately
