### 2020-12-14 #2879 @meg-simula
@arfon Any ideas on why @ramisetti is not assigned to this issue (and not allowed to edit the checkboxes)? Anything I can do regarding this?
