### 2020-12-14 #2879 @mscroggs
Yes, this format is fine. I'll get some revisions done soon
