### 2020-12-14 #2879 @arfon
> @arfon Any ideas on why @ramisetti is not assigned to this issue (and not allowed to edit the checkboxes)? Anything I can do regarding this?

@meg-simula - looks like something went awry with the repository invite. Can you try clicking the above link again @ramisetti? The invite should be shown at the top of this page: https://github.com/openjournals/joss-reviews/invitations
