### 2020-12-14 #2879 @ramisetti
Thank you @arfon and @meg-simula. I can now edit the checkboxes.
