### 2020-12-14 #2879 @arfon
> Thanks @arfon. Next time this happens, should I just try this re-invite?

Yes, that's it!
