### 2021-01-19 119be82
add testing to README. Skip dolfin tests if not installed
