### 2021-01-19 bempp/bempp-cl#94 @mscroggs
JOSS paper
Checklist of things that need doing. These are in progress in the [mscroggs/joss](https://github.com/bempp/bempp-cl/tree/mscroggs/joss) branch.

From [jamtrott's review](https://github.com/bempp/bempp-cl/issues/89):
- [x] (i) "there is lack of comparison to other software for solving boundary integral equations"
- [x] (ii) "I think the paper could benefit from a little more background, such as an example of the type of equations that are relevant."
- [x] (xiv) "The statement of need could benefit from a short discussion or mention of why they choose to use OpenCL and numba for just-in-time compilation."

From [ramisetti's review](https://github.com/openjournals/joss-reviews/issues/2879#issuecomment-743716755):
- [x] (iii) "The documentation claims the software can import and export in a number of formats, including Gmsh and VTK but it is not clearly evident from the examples provided in the applications section, except the exporting of gmsh file in the second example. The authors should provide at least one example for each case."
- [x] (iv) "Similarly, the authors should provide examples on the CPU and GPU parallelisation."
- [x] (v) "The documentation for the examples should place a phrase related to the functionality to help the users understand the purpose of the example. For example, the documentation for each example can mention a phrase about the functionality in brackets beside the link"
- [x] (vi) "The authors may consider mentioning about the target audience of this software to make it more clear to the potential users"
- [x] (vii) "The authors may consider mentioning the versions of the dependencies particularly for gmsh as this could save time for users in installing the dependencies"
- [x] (viii) "The authors should fix the broken link for “Weakly imposing a Dirichlet boundary condition” example."
- [x] (ix) "I would recommend to separate the tutorials from documentation as it is more intuitive to find examples in tutorials section rather in documentation"
- [x] (x) "The python core functionality documentation is available through few links, but I believe it is not up to a satisfactory level to be useful for the prospective software developers interested in the software. The documentation should explicitly provide a link to API documentation"
- [x] (xi) "Although the repository contains a list of unit test but the documentation does not mention anything about any automated tests. The authors should consider updating the documentation with information on unit tests"
- [x] (xii) "Information on third party contribution to the software is not provided."
- [x] (xiii) "This submission lacks information on why this software is better and how this compares to any of the existing open source software. The authors should consider including information in the paper if there are any existing other opensource software and how this software compares with them."

Notes on these changes are in the replies below. Some of these changes will only happen once the branch has been merged
