+++ joss/paper.md
@@ -24,7 +24,7 @@ bibliography: paper.bib
 
 # Summary
 The boundary element method (BEM) is a numerical method for approximating the solution of certain types of partial 
-differential equations (PDEs) in homogeneous bounded or unbounded domains. The method finds the approximation by discretising 
+differential equations (PDEs) in homogeneous bounded or unbounded domains. The method finds an approximation by discretising 
 a boundary integral equation that can be derived from the PDE. The mathematical background of BEM is covered in, for example, 
 @Stein07 or @McLean. Typical applications of BEM include electrostatic problems, and acoustic and electromagnetic scattering.
 
@@ -40,7 +40,7 @@ of efforts to fully move to Python, and is an almost complete rewrite of Bempp.
 Bempp-cl provides a comprehensive collection of routines for the assembly of boundary integral operators to solve a wide
 range of relevant application problems. It contains an operator algebra that allows a straight-forward implementation of
 complex operator preconditioned systems of boundary integral equations [@operatoralg] and in particular implements
-everything that is required for Calder\'on preconditioned Maxwell [@maxwellbempp] problems. Bempp-cl uses PyOpenCL [@pyopencl]
+everything that is required for Calderón preconditioned Maxwell [@maxwellbempp] problems. Bempp-cl uses PyOpenCL [@pyopencl]
 to just-in-time compile its computational kernels on a wide range of CPU and GPU devices and modern architectures. Alternatively,
 a fallback Numba implementation is provided.
 
@@ -86,18 +86,18 @@ and GPU compute devices. On systems without OpenCL support, Numba [@numba] is us
 Python-based assembly kernels, giving a slower but still viable alternative to OpenCL.
 
 Bempp-cl provides an interface to the Exafmm-t library [@exafmm] for faster assembly of larger problems with lower memory
-requirements using the fast multipole method (FMM). The interface to Exafmm-t is writting in a generic way so that other
+requirements using the fast multipole method (FMM). The interface to Exafmm-t is written in a generic way so that other
 FMM libraries or alternative matrix compression techniques could be used in future. 
 
 The submodule `bempp.api.linalg` contains wrapped versions of SciPy's [@scipy] LU, CG, and GMRes solvers. By using 
 SciPy's `LinearOperator` interface, Bempp-cl's boundary operators can easily be used with other iterative solvers.
 
 ## Potential and far field operators
-Potential and far fiels operators for the evaluation at points in the domain or the asymptotic behavior at infinity are 
+Potential and far field operators for the evaluation at points in the domain or the asymptotic behavior at infinity are 
 included in the `bempp.api.operators.potential` and `bempp.api.operators.far_field` submodules.
 
 ## Further information
-Full documentation of the library, including a number of example Jupyter notebooks, can be found online at ``bempp.com`` and in
+Full documentation of the library, including a number of example Jupyter notebooks, can be found online at ``bempp.com`` and
 in the in-development Bempp Handbook [@bempphandbook].
 
 # Acknowledgements
