### 2021-01-19 30ae1ec
schedule tests every Monday morning
