### 2021-01-19 0e73120
also run circleci tests weekly
