+++ .circleci/config.yml
@@ -39,11 +39,12 @@ workflows:
           filters:
             branches:
               only: master
-    nightly:
-      triggers:
-        - schedule:
-          cron: "0 7 * * 1"
-          filters:
-            branches:
-              only:
-                - master
+  weekly:
+    jobs:
+      - large-tests
+    triggers:
+      - schedule:
+        cron: "0 7 * * 1"
+        filters:
+          branches:
+            only: master
