+++ .github/workflows/on-master-push.yml
@@ -2,8 +2,6 @@ on:
   push:
     branches:
       - master
-  schedule:
-    - cron: "0 7 * * 1"
 
 jobs:
   build-and-publish-docker:
