+++ .github/workflows/on-push.yml
@@ -17,7 +17,7 @@ jobs:
               echo ::set-output name=tag::latest
             else
               the_tag = ${GITHUB_REF#refs/*/}
-              echo ::set-output name=tag::${the_tag/\//--}
+              echo ::set-output name=tag::${the_tag/\//__}
             fi
 
   check-version-number:
@@ -46,7 +46,7 @@ jobs:
         uses: actions/cache@v2
         with:
           path: ~/cache/Dockerfile
-          key: ${{ runner.os }}-Dockerfile-${{ needs.get-branch.outputs.tag//\/_}}-${{ hashFiles('Dockerfile') }}
+          key: ${{ runner.os }}-Dockerfile-${{ needs.get-branch.outputs.tag }}-${{ hashFiles('Dockerfile') }}
       - name: touch ~/cache/Dockerfile
         run: touch ~/cache/Dockerfile
       - name: compare Dockerfile with cache
@@ -66,7 +66,7 @@ jobs:
           username: ${{ secrets.DOCKER_USERNAME }}
           password: ${{ secrets.DOCKER_PASSWORD }}
           repository: bempp/cl-dev-env
-          tags: ${{ needs.get-branch.outputs.tag//\/_ }}
+          tags: ${{ needs.get-branch.outputs.tag }}
           target: bempp-dev-env-with-exafmm
           cache_froms: bempp/cl-dev-env
         if: steps.diff.outputs.changed == 'yes'
@@ -101,7 +101,7 @@ jobs:
       - build-and-publish-docker
       - get-branch
     runs-on: ubuntu-latest
-    container: bempp/cl-dev-env:${{ needs.get-branch.outputs.tag//\/_ }}
+    container: bempp/cl-dev-env:${{ needs.get-branch.outputs.tag }}
     steps:
       - uses: actions/checkout@v2
       - run: python3 setup.py install
