+++ .github/workflows/on-push.yml
@@ -16,8 +16,7 @@ jobs:
             then
               echo ::set-output name=tag::latest
             else
-              the_tag = ${GITHUB_REF#refs/*/}
-              echo ::set-output name=tag::${the_tag/\//__}
+              echo ::set-output name=tag::${GITHUB_REF#refs/*/}
             fi
 
   check-version-number:
@@ -46,7 +45,7 @@ jobs:
         uses: actions/cache@v2
         with:
           path: ~/cache/Dockerfile
-          key: ${{ runner.os }}-Dockerfile-${{ needs.get-branch.outputs.tag }}-${{ hashFiles('Dockerfile') }}
+          key: ${{ runner.os }}-Dockerfile-${{needs.get-branch.outputs.tag/\/_}}-${{ hashFiles('Dockerfile') }}
       - name: touch ~/cache/Dockerfile
         run: touch ~/cache/Dockerfile
       - name: compare Dockerfile with cache
@@ -66,7 +65,7 @@ jobs:
           username: ${{ secrets.DOCKER_USERNAME }}
           password: ${{ secrets.DOCKER_PASSWORD }}
           repository: bempp/cl-dev-env
-          tags: ${{ needs.get-branch.outputs.tag }}
+          tags: ${{needs.get-branch.outputs.tag/\/_}}
           target: bempp-dev-env-with-exafmm
           cache_froms: bempp/cl-dev-env
         if: steps.diff.outputs.changed == 'yes'
