### 2021-01-19 bempp/bempp-cl#94 @mscroggs
(ix) and (v) tutorials moved to https://bempp.com/documentation/tutorials.html and one sentence descriptions added
