### 2021-01-19 ddf82f3
update docker script to v2
