+++ .github/workflows/on-push.yml
@@ -60,7 +60,7 @@ jobs:
         run: cp Dockerfile ~/cache/Dockerfile
 
       - name: Build and push testing Docker image with exafmm
-        uses: docker/build-push-action@v1
+        uses: docker/build-push-action@v2
         with:
           username: ${{ secrets.DOCKER_USERNAME }}
           password: ${{ secrets.DOCKER_PASSWORD }}
