+++ .github/workflows/on-push.yml
@@ -16,7 +16,7 @@ jobs:
             then
               echo ::set-output name=tag::latest
             else
-              echo ::set-output name=tag::${GITHUB_REF#refs/*/}
+              echo ::set-output name=tag::${${GITHUB_REF#refs/*/}/\//_}
             fi
 
   check-version-number:
@@ -36,6 +36,8 @@ jobs:
     runs-on: ubuntu-latest
     needs: get-branch
     steps:
+      - name: echo tag
+        uses: echo "${{needs.get-branch.outputs.tag}}"
       - name: Checkout code
         uses: actions/checkout@v2
       - name: Make cache folder
