+++ .github/workflows/on-push.yml
@@ -16,7 +16,7 @@ jobs:
             then
               echo ::set-output name=tag::latest
             else
-              echo ::set-output name=tag::${${GITHUB_REF#refs/*/}/\//_}
+              echo ::set-output name=tag::${GITHUB_REF#refs/*//\//_}
             fi
 
   check-version-number:
@@ -47,7 +47,7 @@ jobs:
         uses: actions/cache@v2
         with:
           path: ~/cache/Dockerfile
-          key: ${{ runner.os }}-Dockerfile-${{needs.get-branch.outputs.tag}}-${{ hashFiles('Dockerfile') }}
+          key: ${{ runner.os }}-Dockerfile-${{ needs.get-branch.outputs.tag }}-${{ hashFiles('Dockerfile') }}
       - name: touch ~/cache/Dockerfile
         run: touch ~/cache/Dockerfile
       - name: compare Dockerfile with cache
