### 2021-01-19 bempp/bempp-cl#94 @mscroggs
(vii) Version number for Gmsh has been added to https://bempp.com/installation.html. All the other dependencies are tested against latest releases
