+++ .github/workflows/on-push.yml
@@ -37,8 +37,6 @@ jobs:
     runs-on: ubuntu-latest
     needs: get-branch
     steps:
-      - name: echo tag
-        run: echo "${{ needs.get-branch.outputs.tag }}"
       - name: Checkout code
         uses: actions/checkout@v2
       - name: Make cache folder
@@ -62,13 +60,17 @@ jobs:
       - name: Copy Dockerfile to ~/cache/Dockerfile
         run: cp Dockerfile ~/cache/Dockerfile
 
-      - name: Build and push testing Docker image with exafmm
-        uses: docker/build-push-action@v2
+      - name: Login to DockerHub
+        uses: docker/login-action@v1
         with:
           username: ${{ secrets.DOCKER_USERNAME }}
           password: ${{ secrets.DOCKER_PASSWORD }}
-          repository: bempp/cl-dev-env
-          tags: ${{needs.get-branch.outputs.tag}}
+
+      - name: Build and push testing Docker image with exafmm
+        uses: docker/build-push-action@v2
+        with:
+          push: true
+          tags: bempp/cl-dev-env:${{ needs.get-branch.outputs.tag }}
           target: bempp-dev-env-with-exafmm
           cache_froms: bempp/cl-dev-env
         if: steps.diff.outputs.changed == 'yes'
