+++ joss/paper.bib
@@ -26,7 +26,7 @@
     doi={10.1002/nme.2579}
 }
 
-@mish{meshio,
+@misc{meshio,
     author={Nico Schlömer and other contributors},
     title={meshio: {I/O} for mesh files},
     url = {https://github.com/nschloe/meshio},
@@ -159,4 +159,31 @@ AUTHOR = {S. M. Rao and D. R. Wilton and A. W. Glisson},
     NUMBER = {2},
      PAGES = {6:1--6:40},
     DOI = {10.1145/2590830}
+}
+
+@article{PyGBe,
+  doi = {10.21105/joss.00043},
+  url = {https://doi.org/10.21105/joss.00043},
+  year = {2016},
+  publisher = {The Open Journal},
+  volume = {1},
+  number = {4},
+  pages = {43},
+  author = {Christopher D. Cooper and Natalia C. Clementi and Gilbert Forsyth and Lorena A. Barba},
+  title = {PyGBe: Python, GPUs and Boundary elements for biomolecular electrostatics},
+  journal = {Journal of Open Source Software}
+}
+
+@techreport{BETL,
+  author = {R. Hiptmair and L. Kielhorn},
+  title = {BETL -- A generic boundary element template library},
+  number = {2012-36},
+  institution = {Seminar for Applied Mathematics, ETH Zürich},
+  url = {ftp://ftp.sam.math.ethz.ch/pub/sam-reports/reports/reports2012/2012-36.pdf}
+}
+
+@misc{abem,
+    author={Frank Jargstorff},
+    title={Acoustic Boundary Element Method (abem)},
+    url = {https://github.com/fjargsto/abem},
 }
