### 2021-01-21 2b6a082
pydocstyle, add some module docstrings
