+++ bempp/api/utils/parameters.py
@@ -8,7 +8,6 @@ class _Verbosity(object):
 
     def __init__(self):
         """Initialize verbosity parameters."""
-
         self.extended_verbosity = False
 
 
@@ -17,7 +16,6 @@ class _Output(object):
 
     def __init__(self):
         """Initialize output parameters."""
-
         self.gmsh_use_binary = True
 
 
@@ -25,7 +23,7 @@ class _Quadrature(object):
     """Quadrature orders."""
 
     def __init__(self):
-
+        """Iniitalize quadrature parameters."""
         self.regular = 4
         self.singular = 4
 
@@ -34,7 +32,7 @@ class _Fmm(object):
     """Fmm options."""
 
     def __init__(self):
-
+        """Iniitalize FMM parameters."""
         self.expansion_order = 5
         self.depth = 4
         self.ncrit = 400
@@ -47,6 +45,7 @@ class _DenseAssembly(object):
     """Dense assembly options."""
 
     def __init__(self):
+        """Iniitalize dense assembly parameters."""
         self.workgroup_size_multiple = 2
 
 
@@ -54,6 +53,7 @@ class _Assembly(object):
     """Assembly options."""
 
     def __init__(self):
+        """Iniitalize assembly parameters."""
         self.dense = _DenseAssembly()
         self.always_promote_to_double = False
         self.discretization_type = "galerkin"
@@ -64,7 +64,6 @@ class DefaultParameters(object):
 
     def __init__(self):
         """Initialize parameters."""
-
         self.verbosity = _Verbosity()
         self.output = _Output()
         self.quadrature = _Quadrature()
