+++ bempp/api/operators/far_field/__init__.py
@@ -1,2 +1,4 @@
+"""Far field operators."""
+
 from . import helmholtz
 from . import maxwell
