+++ bempp/__init__.py
@@ -1,0 +1,8 @@
+"""Bempp: Boundary element method Python package.
+
+Bempp-cl is an open-source boundary element method library that
+can be used to assemble all the standard integral kernels for
+Laplace, Helmholtz, modified Helmholtz, and Maxwell problems.
+
+Documentation can be found at https://bempp.com/documentation/index.html
+"""
