+++ bempp/api/linalg/__init__.py
@@ -1,3 +1,9 @@
+"""Linear solvers.
+
+Documentation of this module can be found at
+https://bempp.com/handbook/api/solvers.html
+"""
+
 from .iterative_solvers import gmres
 from .iterative_solvers import cg
 from .direct_solvers import lu
