+++ bempp/core/numba_kernels.py
@@ -1,3 +1,5 @@
+"""Kernels for assembly using Numba."""
+
 import numba as _numba
 import numpy as _np
 
@@ -90,7 +92,6 @@ def select_numba_kernels(operator_descriptor, mode="regular"):
 )
 def get_piola_transform(grid_data, elements, local_points):
     """Compute the Piola transform."""
-
     npoints = local_points.shape[1]
     nelements = len(elements)
     result = _np.zeros((nelements, 3, 3, npoints), dtype=local_points.dtype)
@@ -111,7 +112,6 @@ def get_piola_transform(grid_data, elements, local_points):
 )
 def get_edge_lengths(grid_data, elements):
     """Compute the edge lengths for the given elements."""
-
     nelements = len(elements)
     result = _np.zeros((nelements, 3), dtype=grid_data.vertices.dtype)
 
@@ -187,7 +187,7 @@ def elements_adjacent(elements, index1, index2):
 def laplace_single_layer_regular(
     test_point, trial_points, test_normal, trial_normals, kernel_parameters
 ):
-    """Laplace single layer for regular kernels."""
+    """Evaluate Laplace single layer for regular kernels."""
     npoints = trial_points.shape[1]
     dtype = trial_points.dtype
     output = _np.zeros(npoints, dtype=dtype)
@@ -206,7 +206,7 @@ def laplace_single_layer_regular(
 def laplace_double_layer_regular(
     test_point, trial_points, test_normal, trial_normals, kernel_parameters
 ):
-    """Laplace double layer for regular kernels."""
+    """Evaluate Laplace double layer for regular kernels."""
     npoints = trial_points.shape[1]
     dtype = trial_points.dtype
     output = _np.zeros(npoints, dtype=dtype)
@@ -233,7 +233,7 @@ def laplace_double_layer_regular(
 def laplace_adjoint_double_layer_regular(
     test_point, trial_points, test_normal, trial_normals, kernel_parameters
 ):
-    """Laplace adjoint double layer for regular kernels."""
+    """Evaluate Laplace adjoint double layer for regular kernels."""
     npoints = trial_points.shape[1]
     dtype = trial_points.dtype
     output = _np.zeros(npoints, dtype=dtype)
@@ -260,7 +260,7 @@ def laplace_adjoint_double_layer_regular(
 def laplace_single_layer_singular(
     test_points, trial_points, test_normal, trial_normal, kernel_parameters
 ):
-    """Laplace single layer for singular kernels."""
+    """Evaluate Laplace single layer for singular kernels."""
     npoints = trial_points.shape[1]
     dtype = trial_points.dtype
     output = _np.zeros(npoints, dtype=dtype)
@@ -279,7 +279,7 @@ def laplace_single_layer_singular(
 def laplace_double_layer_singular(
     test_points, trial_points, test_normal, trial_normal, kernel_parameters
 ):
-    """Laplace double layer for singular kernels."""
+    """Evaluate Laplace double layer for singular kernels."""
     npoints = trial_points.shape[1]
     dtype = trial_points.dtype
     output = _np.zeros(npoints, dtype=dtype)
@@ -306,7 +306,7 @@ def laplace_double_layer_singular(
 def laplace_adjoint_double_layer_singular(
     test_points, trial_points, test_normal, trial_normal, kernel_parameters
 ):
-    """Laplace adjoint double layer for singular kernels."""
+    """Evaluate Laplace adjoint double layer for singular kernels."""
     npoints = trial_points.shape[1]
     dtype = trial_points.dtype
     output = _np.zeros(npoints, dtype=dtype)
[…]
