+++ bempp/api/operators/potential/__init__.py
@@ -1,3 +1,9 @@
+"""Functions for creating potential operators.
+
+Full documentation of the functionality of this module
+can be found at https://bempp.com/handbook/api/potential_operators.html.
+"""
+
 from . import laplace
 from . import helmholtz
 from . import maxwell
