### 2021-01-21 8f76113
link to handbook in docstrings
