+++ bempp/api/grid/__init__.py
@@ -1,7 +1,7 @@
 """The Grid classes.
 
-Documentation of this module can be found at
-https://bempp.com/handbook/api/grids.html
+Documentation of the functionality of this module and examples
+can be found at https://bempp.com/handbook/api/grids.html
 """
 
 from .grid import Grid
