+++ bempp/core/__init__.py
@@ -1,1 +1,5 @@
-"""Core Bempp assembly routines."""
+"""Core Bempp assembly routines.
+
+Documentation of the functionality of this module and examples
+can be found at https://bempp.com/handbook/core/assembling_operators.html
+"""
