+++ bempp/api/operators/boundary/__init__.py
@@ -1,6 +1,6 @@
 """Functions for creating boundary integral operators.
 
-Full documentation of the functionality of this module
+Documentation of the functionality of this module and examples
 can be found at https://bempp.com/handbook/api/boundary_operators.html.
 """
 
