+++ bempp/api/operators/potential/__init__.py
@@ -1,6 +1,6 @@
 """Functions for creating potential operators.
 
-Full documentation of the functionality of this module
+Documentation of the functionality of this module and examples
 can be found at https://bempp.com/handbook/api/potential_operators.html.
 """
 
