+++ bempp/api/space/__init__.py
@@ -1,7 +1,7 @@
 """Definitions of finite element spaces.
 
-Documentation of the functionality available in this
-module can be found at https://bempp.com/handbook/api/function_spaces.html
+Documentation of the functionality of this module and examples
+can be found at https://bempp.com/handbook/api/function_spaces.html
 """
 
 from .space import function_space
