### 2021-01-21 bempp/bempp-cl#94 @mscroggs
(x) More docstrings have been added, and links between [the Read the Docs](https://bempp-cl.readthedocs.io/en/latest/index.html) and [the Handbook](https://bempp.com/handbook/intro.html) have been added
