### 2021-01-21 bempp/bempp-cl#94 @mscroggs
(iv) Snippets showing how to do this have been added to [the Handbook](https://bempp.com/handbook/core/assembling_operators.html#assembling-on-different-devices). Longer term, we are planning to write a tutorial showing off GPU and CPU functionality.
