+++ joss/paper.bib
@@ -120,6 +120,13 @@ AUTHOR = {S. M. Rao and D. R. Wilton and A. W. Glisson},
        url = {https://bempp.com/handbook/},
 }
 
+@misc{bempp-maxwell-example,
+    AUTHOR = {Timo Betcke and Matthew W. Scroggs},
+     TITLE = {Bempp tutorial: Electromagnetic scattering from flat screens},
+      YEAR = {2020},
+       url = {https://nbviewer.jupyter.org/github/bempp/bempp-cl/blob/master/notebooks/maxwell/maxwell_screen.ipynb}
+}
+
 @book {Stein07,
     AUTHOR = {Steinbach, Olaf},
      TITLE = {Numerical approximation methods for elliptic boundary value problems},
