+++ joss/paper.bib
@@ -120,7 +120,7 @@ AUTHOR = {S. M. Rao and D. R. Wilton and A. W. Glisson},
        url = {https://bempp.com/handbook/},
 }
 
-@misc{bempp-maxwell-example,
+@misc{Bempp-maxwell-example,
     AUTHOR = {Timo Betcke and Matthew W. Scroggs},
      TITLE = {Bempp tutorial: Electromagnetic scattering from flat screens},
       YEAR = {2020},
