+++ bempp/core/dispatcher.py
@@ -7,10 +7,12 @@ def singular_assembler_dispatcher(device_interface, *args):
 
     if interface_type == "opencl":
         from bempp.core.opencl_assemblers import singular_assembler
+
         singular_assembler(device_interface, *args)
 
     elif interface_type == "numba":
         from bempp.core.numba_assemblers import singular_assembler
+
         singular_assembler(device_interface, *args)
 
     else:
@@ -23,10 +25,12 @@ def dense_assembler_dispatcher(device_interface, *args):
 
     if interface_type == "opencl":
         from bempp.core.opencl_assemblers import dense_assembler
+
         dense_assembler(device_interface, *args)
 
     elif interface_type == "numba":
         from bempp.core.numba_assemblers import dense_assembler
+
         dense_assembler(device_interface, *args)
 
     else:
@@ -39,10 +43,12 @@ def potential_dispatcher(device_interface, *args):
 
     if interface_type == "numba":
         from bempp.core.numba_assemblers import potential_assembler
+
         return potential_assembler(device_interface, *args)
 
     elif interface_type == "opencl":
         from bempp.core.opencl_assemblers import potential_assembler
+
         return potential_assembler(device_interface, *args)
 
     else:
