+++ .github/scripts/update_version_number.py
@@ -25,7 +25,9 @@ if v2 > version:
     version = v2
 
 vfile3 = bempp.get_contents("bempp/version.py", branch.commit.sha)
-v3 = tuple(int(i) for i in vfile3.decoded_content.split("\n", 1)[1].split(b'"')[1].split(b"."))
+v3 = tuple(
+    int(i) for i in vfile3.decoded_content.split("\n", 1)[1].split(b'"')[1].split(b".")
+)
 if v3 > version:
     version = v3
 
