### 2021-01-21 bempp/bempp-cl#94 @mscroggs
Useful link to diff of md and bib files to see what has changed in paper: https://github.com/bempp/bempp-cl/pull/95/files?file-filters%5B%5D=.bib&file-filters%5B%5D=.md
