### 2021-01-21 bempp/bempp-cl#94 @mscroggs
Information about how to run tests has been added to [readme](https://github.com/bempp/bempp-cl/blob/master/README.md)
