### 2021-01-21 bempp/bempp-cl#94 @mscroggs
[CONTRIBUTING.md](https://github.com/bempp/bempp-cl/blob/master/CONTRIBUTING.md) has been added
