+++ .github/workflows/on-master-push.yml
@@ -10,12 +10,17 @@ jobs:
     steps:
       - name: Checkout code
         uses: actions/checkout@v2
-      - name: Build and push Docker image
-        uses: docker/build-push-action@v1
+
+      - name: Login to DockerHub
+        uses: docker/login-action@v1
         with:
           username: ${{ secrets.DOCKER_USERNAME }}
           password: ${{ secrets.DOCKER_PASSWORD }}
-          repository: bempp/cl-notebook
-          tags: latest
+
+      - name: Build and push testing Docker image with exafmm
+        uses: docker/build-push-action@v2
+        with:
+          push: true
+          tage: bempp/cl-notebook:latest
           target: lab
           cache_froms: bempp/cl-notebook:latest
