### 2021-01-21 #2879 @mscroggs
I've updated the paper following the reviewer comments. Notes on changes made can be found in this issue: https://github.com/bempp/bempp-cl/issues/94
