+++ .github/workflows/on-push.yml
@@ -60,19 +60,19 @@ jobs:
       - name: Copy Dockerfile to ~/cache/Dockerfile
         run: cp Dockerfile ~/cache/Dockerfile
 
+      - name: Set up Docker Buildx
+        uses: docker/setup-buildx-action@v1
       - name: Login to DockerHub
         uses: docker/login-action@v1
         with:
           username: ${{ secrets.DOCKER_USERNAME }}
           password: ${{ secrets.DOCKER_PASSWORD }}
-
       - name: Build and push testing Docker image with exafmm
         uses: docker/build-push-action@v2
         with:
           push: true
           tags: bempp/cl-dev-env:${{ needs.get-branch.outputs.tag }}
           target: bempp-dev-env-with-exafmm
-          cache_froms: bempp/cl-dev-env
         if: steps.diff.outputs.changed == 'yes'
   
   style-checks:
