### 2021-01-21 1b05bf1
remove installing pytest-xdist from latest images
