+++ .github/workflows/on-pr.yml
@@ -34,6 +34,5 @@ jobs:
       - uses: actions/checkout@v2
       - run: python3 setup.py install
         name: Install Bempp
-      - run: pip3 install pytest-xdist
       - run: python3 -m pytest -n4 --durations=50 test/unit
         name: Run Bempp unit tests
