+++ bempp/core/opencl_kernels.py
@@ -7,8 +7,11 @@ _CURRENT_PATH = _os.path.dirname(_os.path.realpath(__file__))
 _INCLUDE_PATH = _os.path.abspath(_os.path.join(_CURRENT_PATH, "./sources/include"))
 _KERNEL_PATH = _os.path.abspath(_os.path.join(_CURRENT_PATH, "./sources/kernels"))
 
-_DEFAULT_DEVICE = None
-_DEFAULT_CONTEXT = None
+_DEFAULT_CPU_DEVICE = None
+_DEFAULT_CPU_CONTEXT = None
+
+_DEFAULT_GPU_DEVICE = None
+_DEFAULT_GPU_CONTEXT = None
 
 
 def select_cl_kernel(operator_descriptor, mode):
@@ -108,7 +111,7 @@ def get_kernel_compile_options(options, precision):
     return compile_options
 
 
-def build_program(assembly_function, options, precision):
+def build_program(assembly_function, options, precision, device_type="cpu"):
     """Build the kernel and return it."""
 
     file_name = assembly_function + ".cl"
@@ -118,22 +121,22 @@ def build_program(assembly_function, options, precision):
     kernel_options = get_kernel_compile_options(options, precision)
 
     return (
-        _cl.Program(default_context(), kernel_string)
+        _cl.Program(default_context(device_type), kernel_string)
         .build(options=kernel_options)
         .kernel_function
     )
 
 
 def get_kernel_from_operator_descriptor(
-    operator_descriptor, options, mode, force_novec=False
+    operator_descriptor, options, mode, force_novec=False, device_type="cpu"
 ):
     """Return compiled kernel from operator descriptor."""
 
     precision = operator_descriptor.precision
     assembly_function, kernel_name = select_cl_kernel(operator_descriptor, mode=mode)
 
-    vec_length = get_vector_width(precision)
-    vec_string = get_vec_string(precision)
+    vec_length = get_vector_width(precision, device_type)
+    vec_string = get_vec_string(precision, device_type)
 
     if not mode == "singular":
         if force_novec or vec_length == 1:
@@ -143,77 +146,139 @@ def get_kernel_from_operator_descriptor(
     options["KERNEL_FUNCTION"] = kernel_name
     options["VEC_LENGTH"] = vec_length
     options["VEC_STRING"] = vec_string
-    return build_program(assembly_function, options, precision)
+    return build_program(assembly_function, options, precision, device_type)
 
 
-def get_kernel_from_name(name, options, precision="double"):
+def get_kernel_from_name(name, options, precision="double", device_type="cpu"):
     """Return compiled kernel from name."""
 
-    vec_length = get_vector_width(precision)
-    vec_string = get_vec_string(precision)
+    vec_length = get_vector_width(precision, device_type)
+    vec_string = get_vec_string(precision, device_type)
 
     options["VEC_LENGTH"] = vec_length
     options["VEC_STRING"] = vec_string
-    return build_program(name, options, precision)
+    return build_program(name, options, precision, device_type)
 
 
-def get_vector_width(precision):
+def get_vector_width(precision, device_type="cpu"):
     """Return vector width."""
     import bempp.api
 
     mode_to_length = {"novec": 1, "vec4": 4, "vec8": 8, "vec16": 16}
 
     if bempp.api.VECTORIZATION_MODE == "auto":
-        return get_native_vector_width(default_device(), precision)
[…]
