### 2021-01-23 5688184
Fixed automatic GPU detection
