### 2021-01-23 7377723
Reset Potential Operator device to CPU
