+++ bempp/api/__init__.py
@@ -287,6 +287,6 @@ DEFAULT_PRECISION = "double"
 VECTORIZATION_MODE = "auto"
 
 BOUNDARY_OPERATOR_DEVICE_TYPE = "cpu"
-POTENTIAL_OPERATOR_DEVICE_TYPE = "gpu"
+POTENTIAL_OPERATOR_DEVICE_TYPE = "cpu"
 
 ALL = -1  # Useful global identifier
