+++ .github/workflows/on-release.yml
@@ -27,15 +27,19 @@ jobs:
         id: tag_name
         run: |
           echo ::set-output name=SOURCE_TAG::${GITHUB_REF#refs/tags/}
-      - name: Build and push Docker image
-        uses: docker/build-push-action@v1
+     - name: Set up Docker Buildx
+        uses: docker/setup-buildx-action@v1
+      - name: Login to DockerHub
+        uses: docker/login-action@v1
         with:
           username: ${{ secrets.DOCKER_USERNAME }}
           password: ${{ secrets.DOCKER_PASSWORD }}
-          repository: bempp/cl-notebook
-          tags: ${{ steps.tag_name.outputs.SOURCE_TAG }}
+      - name: Build and push testing Docker image with exafmm
+        uses: docker/build-push-action@v2
+        with:
+          push: true
+          tags: bempp/cl-notebook:${{ steps.tag_name.outputs.SOURCE_TAG }}
           target: lab
-          cache_froms: bempp/cl-notebook:${{ steps.tag_name.outputs.SOURCE_TAG }}
 
   tweet:
     name: Tweet about latest version
