+++ .github/workflows/weekly.yml
@@ -18,3 +18,37 @@ jobs:
           tags: latest
           target: lab
           cache_froms: bempp/cl-notebook:latest
+
+  build-and-publish-docker-test:
+    name: Build and publish testing environment docker image
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v2
+      - name: Set up Docker Buildx
+        uses: docker/setup-buildx-action@v1
+      - name: Login to DockerHub
+        uses: docker/login-action@v1
+        with:
+          username: ${{ secrets.DOCKER_USERNAME }}
+          password: ${{ secrets.DOCKER_PASSWORD }}
+      - name: Build and push testing Docker image with exafmm
+        uses: docker/build-push-action@v2
+        with:
+          push: true
+          tags: bempp/cl-dev-env:latest
+          target: bempp-dev-env-with-exafmm
+  
+  build-and-test-with-exafmm:
+    name: Build Bempp and run tests
+    needs:
+      - build-and-publish-docker-test
+    runs-on: ubuntu-latest
+    container: bempp/cl-dev-env:latest
+    steps:
+      - uses: actions/checkout@v2
+      - run: python3 setup.py install
+        name: Install Bempp
+      - run: python3 -m pytest -n4 --durations=50 test/unit
+        name: Run Bempp unit tests
+
