### 2021-01-27 274cd36
Added test of fenicsx
