+++ test/unit/external/test_fenicsx.py
@@ -1,0 +1,44 @@
+"""Unit tests for the FEniCSx interface."""
+
+import pytest
+import numpy as np
+from mpi4py import MPI
+
+
+def test_p1_trace(allow_external_skips):
+    """Test the trace of a P1 Dolfin function."""
+    if allow_external_skips:
+        try:
+            import dolfinx
+            import dolfinx.geometry
+        except ImportError:
+            pytest.skip("DOLFIN-X must be installed to run this test")
+    else:
+        import dolfinx
+        import dolfinx.geometry
+    import bempp.api
+    from bempp.api.external.fenicsx import fenics_to_bempp_trace_data
+
+    fenics_mesh = dolfinx.UnitCubeMesh(MPI.COMM_WORLD, 2, 2, 2)
+    fenics_space = dolfinx.FunctionSpace(fenics_mesh, ("CG", 1))
+
+    bempp_space, trace_matrix = fenics_to_bempp_trace_data(fenics_space)
+
+    fenics_coeffs = np.random.rand(fenics_space.dim)
+    bempp_coeffs = trace_matrix @ fenics_coeffs
+
+    fenics_fun = dolfinx.Function(fenics_space)
+    fenics_fun.vector[:] = fenics_coeffs
+    bempp_fun = bempp.api.GridFunction(bempp_space, coefficients=bempp_coeffs)
+
+    tree = dolfinx.geometry.BoundingBoxTree(fenics_mesh, 3)
+
+    for cell in bempp_space.grid.entity_iterator(0):
+        mid = cell.geometry.centroid
+        bempp_val = bempp_fun.evaluate(cell.index, np.array([[1 / 3], [1 / 3]]))
+
+        fenics_cell = dolfinx.geometry.compute_colliding_cells(tree, fenics_mesh, mid)[
+            0
+        ]
+        fenics_val = fenics_fun.eval([mid.T], [fenics_cell])
+        assert np.isclose(bempp_val[0, 0], fenics_val[0])
