+++ test/unit/external/test_fenics.py
@@ -4,12 +4,15 @@ import pytest
 import numpy as np
 
 
-def test_p1_trace():
+def test_p1_trace(allow_external_skips):
     """Test the trace of a P1 Dolfin function."""
-    try:
+    if allow_external_skips:
+        try:
+            import dolfin
+        except ImportError:
+            pytest.skip("DOLFIN must be installed to run this test")
+    else:
         import dolfin
-    except ImportError:
-        pytest.skip("DOLFIN must be installed to run this test")
     import bempp.api
     from bempp.api.external.fenics import fenics_to_bempp_trace_data
 
@@ -35,12 +38,15 @@ def test_p1_trace():
         assert np.allclose(bempp_val.T[0], fenics_val)
 
 
-def test_nc1_trace():
+def test_nc1_trace(allow_external_skips):
     """Test the trace of a (N1curl, 1) Dolfin function."""
-    try:
+    if allow_external_skips:
+        try:
+            import dolfin
+        except ImportError:
+            pytest.skip("DOLFIN must be installed to run this test")
+    else:
         import dolfin
-    except ImportError:
-        pytest.skip("DOLFIN must be installed to run this test")
     import bempp.api
     from bempp.api.external.fenics import fenics_to_bempp_trace_data
 
