### 2021-01-27 1235133
don't allow skipping when libraries are missing when running CI
