+++ bempp/api/external/fenicsx.py
@@ -33,7 +33,7 @@ def boundary_grid_from_fenics_mesh(fenics_mesh):
 
 
 def fenics_to_bempp_trace_data(fenics_space):
-    """Returns tuple (space,trace_matrix)."""
+    """Return tuple (space,trace_matrix)."""
     family, degree = fenics_space_info(fenics_space)
 
     if family == "Lagrange":
@@ -44,7 +44,7 @@ def fenics_to_bempp_trace_data(fenics_space):
 
 
 def fenics_space_info(fenics_space):
-    """Returns tuple (family,degree) containing information about a FEniCS space."""
+    """Return tuple (family,degree) containing information about a FEniCS space."""
     element = fenics_space.ufl_element()
     family = element.family()
     degree = element.degree()
