+++ .github/workflows/weekly.yml
@@ -1,3 +1,5 @@
+name: Run weekly tests
+
 on:
   schedule:
     - cron: "0 7 * * 1"
