+++ .github/workflows/on-push.yml
@@ -1,3 +1,5 @@
+name: Run tests and style checks
+
 on: push
 
 jobs:
