+++ .github/workflows/on-push.yml
@@ -64,11 +64,13 @@ jobs:
 
       - name: Set up Docker Buildx
         uses: docker/setup-buildx-action@v1
+        if: steps.diff.outputs.changed == 'yes'
       - name: Login to DockerHub
         uses: docker/login-action@v1
         with:
           username: ${{ secrets.DOCKER_USERNAME }}
           password: ${{ secrets.DOCKER_PASSWORD }}
+        if: steps.diff.outputs.changed == 'yes'
       - name: Build and push testing Docker image with exafmm
         uses: docker/build-push-action@v2
         with:
