+++ test/conftest.py
@@ -28,13 +28,24 @@ def pytest_addoption(parser):
         default="auto",
         help="Valid values: numba opencl",
     )
-    parser.addoption("--allow-external-skips",
-                     help="Set this to \"False\" to prevent skipping tests using external libraries.")
+    parser.addoption("--has-dolfin", type="int" default=0)
+    parser.addoption("--has-dolfinx", type="int" default=0)
+    parser.addoption("--has-exafmm", type="int" default=0)
 
 
 @pytest.fixture
-def allow_external_skips(request):
-    return request.config.getoption("--allow-external-skips") != "False"
+def has_dolfin(request):
+    return request.config.getoption("--has-dolfin") > 0
+
+
+@pytest.fixture
+def has_dolfinx(request):
+    return request.config.getoption("--has-dolfinx") > 0
+
+
+@pytest.fixture
+def has_exafmm(request):
+    return request.config.getoption("--has-exafmm") > 0
 
 
 @pytest.fixture()
