+++ test/unit/external/test_fenicsx.py
@@ -5,17 +5,17 @@ import numpy as np
 from mpi4py import MPI
 
 
-def test_p1_trace(allow_external_skips):
+def test_p1_trace(has_dolfinx):
     """Test the trace of a P1 Dolfin function."""
-    if allow_external_skips:
+    if has_dolfinx:
+        import dolfinx
+        import dolfinx.geometry
+    else:
         try:
             import dolfinx
             import dolfinx.geometry
         except ImportError:
             pytest.skip("DOLFIN-X must be installed to run this test")
-    else:
-        import dolfinx
-        import dolfinx.geometry
     import bempp.api
     from bempp.api.external.fenicsx import fenics_to_bempp_trace_data
 
