+++ .github/workflows/weekly.yml
@@ -27,6 +27,7 @@ jobs:
     steps:
       - name: Checkout code
         uses: actions/checkout@v2
+
       - name: Set up Docker Buildx
         uses: docker/setup-buildx-action@v1
       - name: Login to DockerHub
@@ -38,19 +39,40 @@ jobs:
         uses: docker/build-push-action@v2
         with:
           push: true
+          tags: bempp/cl-dev-env-with-dolfin:latest
+          target: bempp-dev-env-with-dolfin
+        with:
+          push: true
+          tags: bempp/cl-dev-env-with-dolfinx:latest
+          target: bempp-dev-env-with-dolfinx
+        with:
+          push: true
           tags: bempp/cl-dev-env:latest
-          target: bempp-dev-env-with-exafmm
+          target: bempp-dev-env
   
-  build-and-test-with-exafmm:
+  build-and-test-with-dolfin:
     name: Build Bempp and run tests
     needs:
       - build-and-publish-docker-test
     runs-on: ubuntu-latest
-    container: bempp/cl-dev-env:latest
+    container: bempp/cl-dev-env-with-dolfin:latest
     steps:
       - uses: actions/checkout@v2
       - run: python3 setup.py install
         name: Install Bempp
-      - run: python3 -m pytest -n4 --durations=50 test/unit --allow-external-skips False
+      - run: python3 -m pytest -n4 --durations=50 test/unit --has-dolfin 1 --has-dolfinx 0 --has-exafmm 1
+        name: Run Bempp unit tests
+
+  build-and-test-with-dolfinx:
+    name: Build Bempp and run tests
+    needs:
+      - build-and-publish-docker-test
+    runs-on: ubuntu-latest
+    container: bempp/cl-dev-env-with-dolfinx:latest
+    steps:
+      - uses: actions/checkout@v2
+      - run: python3 setup.py install
+        name: Install Bempp
+      - run: python3 -m pytest -n4 --durations=50 test/unit --has-dolfin 0 --has-dolfinx 1 --has-exafmm 1
         name: Run Bempp unit tests
 
