+++ .github/workflows/weekly.yml
@@ -35,16 +35,20 @@ jobs:
         with:
           username: ${{ secrets.DOCKER_USERNAME }}
           password: ${{ secrets.DOCKER_PASSWORD }}
-      - name: Build and push testing Docker image with exafmm
+      - name: Build and push testing Docker image with dolfin
         uses: docker/build-push-action@v2
         with:
           push: true
           tags: bempp/cl-dev-env-with-dolfin:latest
           target: bempp-dev-env-with-dolfin
+      - name: Build and push testing Docker image with dolfinx
+        uses: docker/build-push-action@v2
         with:
           push: true
           tags: bempp/cl-dev-env-with-dolfinx:latest
           target: bempp-dev-env-with-dolfinx
+      - name: Build and push testing Docker image
+        uses: docker/build-push-action@v2
         with:
           push: true
           tags: bempp/cl-dev-env:latest
