+++ .github/workflows/on-push.yml
@@ -71,18 +71,22 @@ jobs:
           username: ${{ secrets.DOCKER_USERNAME }}
           password: ${{ secrets.DOCKER_PASSWORD }}
         if: steps.diff.outputs.changed == 'yes'
-      - name: Build and push testing Docker image with exafmm
+      - name: Build and push testing Docker image with dolfin
         uses: docker/build-push-action@v2
         with:
           push: true
           tags: bempp/cl-dev-env-with-dolfin:${{ needs.get-branch.outputs.tag }}
           target: bempp-dev-env-with-dolfin
         if: steps.diff.outputs.changed == 'yes'
+      - name: Build and push testing Docker image with dolfinx
+        uses: docker/build-push-action@v2
         with:
           push: true
           tags: bempp/cl-dev-env-with-dolfinx:${{ needs.get-branch.outputs.tag }}
           target: bempp-dev-env-with-dolfinx
         if: steps.diff.outputs.changed == 'yes'
+      - name: Build and push testing Docker image
+        uses: docker/build-push-action@v2
         with:
           push: true
           tags: bempp/cl-dev-env:${{ needs.get-branch.outputs.tag }}
