+++ test/unit/external/test_fenicsx.py
@@ -37,8 +37,7 @@ def test_p1_trace(has_dolfinx):
         mid = cell.geometry.centroid
         bempp_val = bempp_fun.evaluate(cell.index, np.array([[1 / 3], [1 / 3]]))
 
-        fenics_cell = dolfinx.geometry.compute_colliding_cells(tree, fenics_mesh, mid)[
-            0
-        ]
+        fenics_cell = dolfinx.geometry.compute_closest_entity(
+            tree, mid, fenics_mesh)[0]
         fenics_val = fenics_fun.eval([mid.T], [fenics_cell])
         assert np.isclose(bempp_val[0, 0], fenics_val[0])
