### 2021-01-27 6bf3592
run dolfin notebooks in the right images
