+++ notebooks/test_notebooks.py
@@ -16,7 +16,12 @@ for dir in ["laplace", "helmholtz", "maxwell"]:
 
 
 @pytest.mark.parametrize(("path", "notebook"), notebooks)
-def test_notebook(path, notebook):
+def test_notebook(path, notebook, has_dolfin, has_dolfinx):
+    if not has_dolfin and notebook.endswith("dolfin.ipynb"):
+        pytest.skip()
+    if not has_dolfinx and notebook.endswith("dolfinx.ipynb"):
+        pytest.skip()
+
     with open(os.path.join(path, notebook)) as f:
         nb = nbformat.read(f, as_version=4)
 
