+++ .circleci/config.yml
@@ -7,11 +7,6 @@ jobs:
     steps:
       - checkout
       - run:
-          command: |
-              apt-get -yq update
-              apt-get -y install jupyter
-          name: Install pytest and jupyter
-      - run:
           command: python3 setup.py install
           name: Install Bempp
       - run:
@@ -20,7 +15,6 @@ jobs:
       - run:
           command: python3 -m pytest -n2 --durations=10 test/validation --has-dolfin 0 --has-dolfinx 0 --has-exafmm 1
           name: Run Bempp validation tests
-
   run-notebooks:
     docker:
       - image: bempp/cl-dev-env:latest
@@ -30,7 +24,7 @@ jobs:
           command: |
               apt-get -yq update
               apt-get -y install jupyter
-          name: Install pytest and jupyter
+          name: Install jupyter
       - run:
           command: python3 setup.py install
           name: Install Bempp
@@ -47,7 +41,7 @@ jobs:
           command: |
               apt-get -yq update
               apt-get -y install jupyter
-          name: Install pytest and jupyter
+          name: Install jupyter
       - run:
           command: python3 setup.py install
           name: Install Bempp
@@ -64,7 +58,7 @@ jobs:
           command: |
               apt-get -yq update
               apt-get -y install jupyter
-          name: Install pytest and jupyter
+          name: Install jupyter
       - run:
           command: python3 setup.py install
           name: Install Bempp
@@ -72,7 +66,6 @@ jobs:
       - run:
           command: python3 -m pytest -n2 notebooks/test_notebooks.py --has-dolfin 0 --has-dolfinx 1 --has-exafmm 1 --dolfin-books-only 1
           name: Run example notebooks
-
   run-benchmarks:
     docker:
       - image: bempp/cl-dev-env:latest
@@ -84,6 +77,9 @@ workflows:
   main:
     jobs:
       - large-tests
+      - run-notebooks
+      - run-notebooks-dolfin
+      - run-notebooks-dolfinx
       - run-benchmarks:
           filters:
             branches:
@@ -91,6 +87,9 @@ workflows:
   weekly:
     jobs:
       - large-tests
+      - run-notebooks
+      - run-notebooks-dolfin
+      - run-notebooks-dolfinx
     triggers:
       - schedule:
           cron: "0 7 * * 1"
