+++ .circleci/config.yml
@@ -29,7 +29,6 @@ jobs:
           command: python3 setup.py install
           name: Install Bempp
       - run:
-      - run:
           command: python3 -m pytest -n2 --durations=10 notebooks/test_notebooks.py --has-dolfin 0 --has-dolfinx 0 --has-exafmm 1
           name: Run example notebooks
   run-notebooks-dolfin:
@@ -46,7 +45,6 @@ jobs:
           command: python3 setup.py install
           name: Install Bempp
       - run:
-      - run:
           command: python3 -m pytest -n2 notebooks/test_notebooks.py --has-dolfin 1 --has-dolfinx 0 --has-exafmm 1 --dolfin-books-only 1
           name: Run example notebooks
   run-notebooks-dolfinx:
@@ -62,7 +60,6 @@ jobs:
       - run:
           command: python3 setup.py install
           name: Install Bempp
-      - run:
       - run:
           command: python3 -m pytest -n2 notebooks/test_notebooks.py --has-dolfin 0 --has-dolfinx 1 --has-exafmm 1 --dolfin-books-only 1
           name: Run example notebooks
