### 2021-01-27 ec1a1e1
remove FEniCSx coupling
