### 2021-01-28 bempp/bempp-cl#94 @mscroggs
CPU/GPU notebook has been added: https://nbviewer.jupyter.org/github/bempp/bempp-cl/blob/master/notebooks/other/opencl_benchmark.ipynb
