### 2021-01-28 #2879 @mscroggs
@ramisetti @jamtrott Thanks for the reviews. I've just finished making changes to the paper and documentation (summary of changes at bempp/bempp-cl#94)
