### 2021-01-28 9522468
add test that shapes can be created
