### 2021-01-28 1467891
Rename CI step. (old) FEniCS no longer included in lab Docker image. Delete test files from lab notebook folder
