### 2021-01-28 c739b7f
run notebook test on other notebooks (but skip cpu/gpu benchmarks)
