### 2021-01-28 5c051ab
Added to_dense and to_sparse methods
