### 2021-01-28 8e2f12a
Replace .A with .to_dense() and .to_sparse()
