+++ test/validation/operators/boundary/test_sparse.py
@@ -30,8 +30,7 @@ def test_sparse_identity_p0_p0(
             device_interface=device_interface,
             precision=precision,
         )
-        .weak_form()
-        .A.todense()
+        .weak_form().to_dense()
     )
 
     _np.testing.assert_allclose(actual, expected, helpers.default_tolerance(precision))
@@ -60,7 +59,7 @@ def test_sparse_identity_p0_p1(
             precision=precision,
         )
         .weak_form()
-        .A.todense()
+        .to_dense()
     )
 
     _np.testing.assert_allclose(actual, expected, helpers.default_tolerance(precision))
@@ -89,7 +88,7 @@ def test_sparse_identity_p1_p0(
             precision=precision,
         )
         .weak_form()
-        .A.todense()
+        .to_dense()
     )
 
     _np.testing.assert_allclose(actual, expected, helpers.default_tolerance(precision))
@@ -117,7 +116,7 @@ def test_sparse_identity_p1_p1(
             precision=precision,
         )
         .weak_form()
-        .A.todense()
+        .to_dense()
     )
 
     _np.testing.assert_allclose(actual, expected, helpers.default_tolerance(precision))
@@ -146,7 +145,7 @@ def test_sparse_identity_snc_rwg(
             precision=precision,
         )
         .weak_form()
-        .A.todense()
+        .to_dense()
     )
 
     if precision == "single":
@@ -180,7 +179,7 @@ def test_sparse_identity_snc_bc(
             precision=precision,
         )
         .weak_form()
-        .A.todense()
+        .to_dense()
     )
 
     if precision == "single":
