+++ test/validation/operators/boundary/test_maxwell_boundary.py
@@ -40,7 +40,7 @@ def test_maxwell_electric_field_sphere(
         atol = 1e-14
 
     expected = helpers.load_npy_data("maxwell_electric_field_boundary")
-    _np.testing.assert_allclose(discrete_op.A, expected, rtol=rtol, atol=atol)
+    _np.testing.assert_allclose(discrete_op.to_dense(), expected, rtol=rtol, atol=atol)
 
 
 def test_maxwell_electric_field_rbc_bc_sphere(
@@ -139,76 +139,6 @@ def test_maxwell_electric_field_bc_sphere(
     bempp.api.clear_fmm_cache()
 
 
-# def test_maxwell_electric_field_complex_sphere(
-#     default_parameters, helpers, device_interface, precision
-# ):
-#     """Test Maxwell electric field on sphere with complex wavenumber."""
-#     from bempp.api import get_precision
-#     from bempp.api import function_space
-#     from bempp.api.operators.boundary.maxwell import electric_field
-
-#     grid = helpers.load_grid("sphere")
-
-#     space1 = function_space(grid, "RWG", 0)
-#     space2 = function_space(grid, "SNC", 0)
-
-#     discrete_op = electric_field(
-#         space1,
-#         space1,
-#         space2,
-#         2.5 + 1j,
-#         assembler="dense",
-#         device_interface=device_interface,
-#         precision=precision,
-#         parameters=default_parameters,
-#     ).weak_form()
-
-#     if precision == "single":
-#         rtol = 1e-5
-#         atol = 1e-7
-#     else:
-#         rtol = 1e-10
-#         atol = 1e-14
-
-#     expected = helpers.load_npy_data("maxwell_electric_field_complex_boundary")
-#     _np.testing.assert_allclose(discrete_op.A, expected, rtol=rtol, atol=atol)
-
-
-# def test_maxwell_electric_field_screen(
-#     default_parameters, helpers, device_interface, precision
-# ):
-#     """Test Maxwell electric field on sphere."""
-#     from bempp.api import get_precision
-#     from bempp.api import function_space
-#     from bempp.api.operators.boundary.maxwell import electric_field
-
-#     grid = helpers.load_grid("structured_grid")
-
-#     space1 = function_space(grid, "RWG", 0)
-#     space2 = function_space(grid, "SNC", 0)
-
-#     discrete_op = electric_field(
-#         space1,
-#         space1,
-#         space2,
-#         2.5,
-#         assembler="dense",
-#         device_interface=device_interface,
-#         precision=precision,
-#         parameters=default_parameters,
-#     ).weak_form()
-
-#     if precision == "single":
-#         rtol = 1e-5
-#         atol = 1e-7
-#     else:
-#         rtol = 1e-10
-#         atol = 1e-14
-
-#     expected = helpers.load_npy_data("maxwell_electric_field_structured_boundary")
-#     _np.testing.assert_allclose(discrete_op.A, expected, rtol=rtol, atol=atol)
-
-
 def test_maxwell_magnetic_field_sphere(
     default_parameters, helpers, device_interface, precision
 ):
@@ -241,701 +171,4 @@ def test_maxwell_magnetic_field_sphere(
 
     expected = helpers.load_npy_data("maxwell_magnetic_field_boundary")
 
-    _np.testing.assert_allclose(discrete_op.A, expected, rtol=rtol, atol=atol)
-
-
-# def test_maxwell_magnetic_field_complex_sphere(
-#     default_parameters, helpers, device_interface, precision
-# ):
[…]
