### 2021-01-28 0dd82e3
add get_diagonal method to diagonal operator
