### 2021-01-28 dab8f2b
only run tests after style checks pass
