### 2021-02-01 #2879 @meg-simula
Thanks @mscroggs, well done!

So, it looks like all the boxes are checked, except one regarding State of the art. @ramisetti is this still an issue?

@jamtrott @ramisetti Are you satisfied with the revisions and changes made by @mscroggs, and have all your points been addressed?
