### 2021-02-04 e9da5c0
Added better routines to select device.
