### 2021-02-05 #2879 @ramisetti
@meg-simula  I am satisfied with the revisions made by @mscroggs. I believe the State of the art may provide a better comparison of the software to other existing ones. The authors may mention a phrase on how Bempp-cl is better/different from other any existing ones.
