### 2021-02-06 bbac208
Updated gitignore and fixed flake8 error
