### 2021-02-07 75b43cb
Vector width now always 1 on GPUs.
