### 2021-02-10 7abf154
Make A return sparse if the matrix is sparse
