+++ notebooks/test_notebooks.py
@@ -16,7 +16,7 @@ for dir in ["laplace", "helmholtz", "maxwell", "other"]:
 def test_notebook(path, notebook, has_dolfin, has_dolfinx, dolfin_books_only):
     # Notebooks in this list will be skipped as the problems are very large or require GPUs
     if notebook in ["reentrant_cube_capacity.ipynb", "opencl_benchmark.ipynb",
-                    "laplace-dirichlet_weak_imposition.ipynb"]:
+                    "dirichlet_weak_imposition.ipynb"]:
         pytest.skip()
 
     if not has_dolfin and notebook.endswith("dolfin.ipynb"):
