### 2021-02-16 a60895e
Add to_sparse and to_dense methods to discrete operators (#102)
* Added to_dense and to_sparse methods

* Replace .A with .to_dense() and .to_sparse()

* add get_diagonal method to diagonal operator

* only run tests after style checks pass

* black formatting

* use to_sparse, not A in fem-bem examples

* add A back in with deprecation warning

* black formatting

* Make A return sparse if the matrix is sparse

* skip weak imposition notebook (it's timing out)

* don't run multiple validation tests at once

* in

* correct notebook name

* skip maxwell fmm tests to reduce chance of timeout

*
[…]
