+++ bempp/api/assembly/discrete_boundary_operator.py
@@ -1,5 +1,6 @@
 """Data structures for assembled boundary operators."""
 
+import warnings
 import numpy as _np
 from bempp.helpers import timeit as _timeit
 from scipy.sparse.linalg.interface import LinearOperator as _LinearOperator
@@ -50,6 +51,24 @@ class _DiscreteOperatorBase(_LinearOperator):
         else:
             return NotImplemented
 
+    @property
+    def A(self):
+        """Return dense matrix."""
+        warnings.warn(
+            "operator.A is deprecated and will be removed in a "
+            "future version. Use operator.to_dense() instead.",
+            DeprecationWarning,
+        )
+        return self.to_dense()
+
+    def to_dense(self):
+        """Return dense matrix."""
+        raise NotImplementedError()
+
+    def to_sparse(self):
+        """Return sparse matrix if operator is sparse."""
+        raise NotImplementedError()
+
 
 class _ScaledDiscreteOperator(_DiscreteOperatorBase):
     """Return a scaled operator."""
@@ -64,10 +83,13 @@ class _ScaledDiscreteOperator(_DiscreteOperatorBase):
         """Matvec."""
         return self._alpha * (self._op @ x)
 
-    @property
-    def A(self):
-        """Return matrix."""
-        return self._alpha * self._op.A
+    def to_dense(self):
+        """Return dense matrix."""
+        return self._alpha * self._op.to_dense()
+
+    def to_sparse(self):
+        """Return sparse matrix if operator is sparse."""
+        return self._alpha * self._op.to_sparse()
 
 
 class _SumDiscreteOperator(_DiscreteOperatorBase):
@@ -91,12 +113,13 @@ class _SumDiscreteOperator(_DiscreteOperatorBase):
         """Evaluate matvec."""
         return self._op1 @ x + self._op2 @ x
 
-    @property
-    def A(self):
-        """Return matrix representation."""
-        res1, res2 = _get_dense(self._op1.A, self._op2.A)
+    def to_dense(self):
+        """Return dense matrix."""
+        return self._op1.to_dense() + self._op2.to_dense()
 
-        return res1 + res2
+    def to_sparse(self):
+        """Return sparse matrix if operator is sparse."""
+        return self._op1.to_sparse() + self._op2.to_sparse()
 
 
 class _ProductDiscreteOperator(_DiscreteOperatorBase):
@@ -120,12 +143,13 @@ class _ProductDiscreteOperator(_DiscreteOperatorBase):
         """Evaluate matvec."""
         return self._op1 @ (self._op2 @ x)
 
-    @property
-    def A(self):
-        """Return matrix representation."""
-        res1, res2 = _get_dense(self._op1.A, self._op2.A)
+    def to_dense(self):
+        """Return dense matrix."""
+        return self._op1.to_dense() @ self._op2.to_dense()
 
-        return res1 @ res2
+    def to_sparse(self):
+        """Return sparse matrix if operator is sparse."""
+        return self._op1.to_sparse() @ self._op2.to_sparse()
 
 
 class GenericDiscreteBoundaryOperator(_DiscreteOperatorBase):
@@ -148,10 +172,13 @@ class GenericDiscreteBoundaryOperator(_DiscreteOperatorBase):
         else:
             return self._evaluator.matvec(x)
 
-    @property
-    def A(self):
-        """Convert to dense."""
+    def to_dense(self):
+        """Return dense matrix."""
         return self @ _np.eye(self.shape[1])
+
+    def to_sparse(self):
+        """Return sparse matrix if operator is sparse."""
+        raise ValueError("This matrix is not sparse.")
 
 
[…]
