### 2021-02-18 #2879 @jamtrott
Hi!

Well done, @mscroggs. I'm happy with the changes, and I think the updated version is a nice improvement.

I don't have any further issues, but I will point out a couple of minor nitpicks.

In the section "Statement of need":
  - "The syntax of library" should perhaps be "The syntax of the library", or maybe "The syntax used by the library", or something.
  -  "[...], making implementing a problem [...]" sounds a bit awkward, and could perhaps be rephrased.
  - In the next paragraph, I find myself a bit confused by the use of parentheses. In particular, "[...] and (abem for acoustics) (Jargstorff, n.d.)", and "PyGBe (for biomolecular electrostatics) (Cooper et al., 2016)". These look a little strange to me, and perhaps it reads better if the parentheses are simply removed.

Once more, good work.

Sorry, it took a while for me to respond.
