### 2021-02-18 c72490e
Update JOSS paper
