+++ joss/paper.bib
@@ -194,3 +194,19 @@ AUTHOR = {S. M. Rao and D. R. Wilton and A. W. Glisson},
     title={Acoustic Boundary Element Method (abem)},
     url = {https://github.com/fjargsto/abem},
 }
+
+@misc{fembemexample,
+    title={Simple {FEM}-{BEM} Coupling for the {H}elmholtz Equation with {FE}ni{CS}x},
+    author={Timo Betcke and Matthew W. Scroggs},
+    note={\url{https://nbviewer.jupyter.org/github/bempp/bempp-cl/blob/master/notebooks/helmholtz/simple_helmholtz_fem_bem_coupling_dolfinx.ipynb}},
+    year={{2021 (accessed 18 February 2021)}}
+}
+
+@book{fenicsbook,
+  title = {Automated Solution of Differential Equations by the Finite Element Method},
+  author = {Anders Logg and Kent-Andre Mardal and Garth N. Wells and others},
+  year = {2012},
+  publisher = {Springer},
+  doi = {10.1007/978-3-642-23099-8},
+  isbn = {978-3-642-23098-1},
+}
