### 2021-02-18 301f765
Update JOSS paper (#105)
* Update JOSS paper

* remove paragraph break
