### 2021-02-18 #2879 @mscroggs
@jamtrott @ramisetti I've fixed those issues and expanded the statement of need a little. You can view a diff of the changes here: bempp/bempp-cl#105
