### 2021-03-15 442a8b1
turn on debug, build dolfinx first
