### 2021-03-16 4959d90
Correct Basix install in docker image (#106)
* change basix build

* no sudo

* try using ninja

* && \

* folders

* turn on debug, build dolfinx first

* =ON

* Release
