+++ Dockerfile
@@ -225,11 +225,10 @@ RUN pip3 install --no-cache-dir meshio>=4.0.16 && \
 # Install Basix
 RUN git clone https://github.com/FEniCS/basix.git basix-src && \
     cd basix-src && \
-    cmake -B build -S . && \
-    cmake --build build && \
-    cmake --install build && \
-    cd python && \
-    pip3 install .
+    cmake -G Ninja -DCMAKE_BUILD_TYPE=Release -DXTENSOR_ENABLE_ASSERT=ON -B build-dir -S . && \
+    cmake --build build-dir && \
+    cmake --install build-dir && \
+    pip3 install ./python
 
 # Install FEniCSx components
 RUN pip3 install --no-cache-dir ipython && \
