### 2021-03-17 6b792bf
fix BOUNDARY_OPERATOR_DEVICE_TYPE error message
The error should now properly show the selected boundary operator device, instead of showing the potential operator device
