### 2021-03-18 #2879 @meg-simula
Thanks again @jamtrott and @ramisetti for completing the reviews - much appreciated - and @mscroggs for addressing all comments. This looks good to go from my side. (The paper might be a bit long for the current JOSS standard, but I think the manuscript is valuable as is, and would recommend its current form.)
