### 2021-03-18 #2879 @meg-simula
PS: whedon check references highlighted one possibly erraneous DOI in the references, would you @mscroggs check that prior to the above, please?
