### 2021-03-18 #2879 @mscroggs
The missing DOI is for an old book that I couldn't find a DOI for. The suggested DOI is for a review of it, not the book itself. I'll sort those tasks out now
