+++ bempp/version.py
@@ -1,3 +1,3 @@
 """Set the version number."""
 
-__version__ = "0.2.2"
+__version__ = "0.2.3"
