### 2021-03-18 #2879 @meg-simula
Ok, sounds good.
--
Marie

On Thu, 18 Mar 2021 at 19:16, Matthew Scroggs ***@***.***>
wrote:

> The missing DOI is for an old book that I couldn't find a DOI for. The
> suggested DOI is for a review of it, not the book itself. I'll sort those
> tasks out now
>
